pyproject.toml
[project]
name = "final-cologne-green-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "duckdb>=1.4.2",
    "folium>=0.20.0",
    "geopandas>=1.1.1",
    "google-cloud-storage>=3.6.0",
    "huggingface_hub>=0.20.0",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.2.1",
    "rasterio>=1.4.3",
    "shapely>=2.1.2",
    "streamlit>=1.51.0",
    "streamlit-folium>=0.25.3",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[tool.uv.workspace]
